fix(core): split georef.rs's compound-plane-angle helpers into georef_parse.rs
The two new raw-byte -0 recovery helpers (plus compound_plane_angle_to_degrees,
moved alongside them for cohesion) pushed rust/core/src/georef.rs to 872
lines against its 758-line module-size ratchet budget
(rust/processing/tests/module_size_ratchet.rs), which is a separate gate
from the TS-only check-module-size.mjs and was red on PR #3591.

Moved compound_plane_angle_to_degrees, compound_angle_has_literal_negative_zero,
and nth_top_level_attribute into a new sibling module georef_parse.rs
(#[path = "georef_parse.rs"] mod georef_parse;, matching the georef_tests.rs
convention already in this file), exposed as pub(super) so georef.rs's
extract_from_site can still call them directly. No behavior change.

georef.rs: 755 lines (was 872, budget 758). georef_parse.rs: 134 lines (new,
under the 400-line new-file limit).
